+++ Realm/RLMRealm.mm
@@ -29,6 +29,8 @@
 #import "RLMUtil.hpp"
 
 #include <exception>
+#include <sys/types.h>
+#include <sys/sysctl.h>
 
 #include <tightdb/version.hpp>
 #include <tightdb/group_shared.hpp>
@@ -86,6 +88,7 @@ NSError *make_realm_error(RLMError code, exception &ex) {
 // Global RLMRealm instance cache
 //
 NSMutableDictionary *s_realmsPerPath;
+NSMutableDictionary *s_keysPerPath;
 
 // FIXME: In the following 3 functions, we should be identifying files by the inode,device number pair
 //  rather than by the path (since the path is not a reliable identifier). This requires additional support
@@ -121,6 +124,9 @@ void clearRealmCache() {
         }
         s_realmsPerPath = [NSMutableDictionary dictionary];
     }
+    @synchronized(s_keysPerPath) {
+        s_keysPerPath = [NSMutableDictionary dictionary];
+    }
 }
 
 static NSString *s_defaultRealmPath = nil;
@@ -145,6 +151,25 @@ void createTablesInTransaction(RLMRealm *realm, RLMSchema *targetSchema) {
     [realm commitWriteTransaction];
 }
 
+bool isDebuggerAttached() {
+    int name[] = {
+        CTL_KERN,
+        KERN_PROC,
+        KERN_PROC_PID,
+        getpid()
+    };
+
+    struct kinfo_proc info;
+    size_t info_size = sizeof(info);
+    if (sysctl(name, sizeof(name)/sizeof(name[0]), &info, &info_size, NULL, 0) == -1) {
+        NSLog(@"sysctl() failed: %s", strerror(errno));
+        return false;
+    }
+
+
+    return (info.kp_proc.p_flag & P_TRACED) != 0;
+}
+
 } // anonymous namespace
 
 NSString * const c_defaultRealmFileName = @"default.realm";
@@ -154,7 +179,6 @@ NSString * const c_defaultRealmFileName = @"default.realm";
     NSThread *_thread;
     NSMapTable *_notificationHandlers;
 
-    std::unique_ptr<LangBindHelper::TransactLogRegistry> _writeLogs;
     std::unique_ptr<Replication> _replication;
     std::unique_ptr<SharedGroup> _sharedGroup;
 
@@ -185,7 +209,19 @@ NSString * const c_defaultRealmFileName = @"default.realm";
     clearRealmCache();
 }
 
-- (instancetype)initWithPath:(NSString *)path readOnly:(BOOL)readonly inMemory:(BOOL)inMemory error:(NSError **)error {
+- (instancetype)initWithPath:(NSString *)path key:(NSData *)key readOnly:(BOOL)readonly inMemory:(BOOL)inMemory error:(NSError **)error {
+    if (key && [key length] != 64) {
+        @throw [NSException exceptionWithName:@"RLMException"
+                                       reason:@"Encryption key must be exactly 64 bytes long"
+                                     userInfo:nil];
+    }
+
+    if (key && isDebuggerAttached()) {
+        @throw [NSException exceptionWithName:@"RLMException"
+                                       reason:@"Cannot open an encrypted Realm with a debugger attached to the process"
+                                     userInfo:nil];
+    }
+
     self = [super init];
     if (self) {
         _path = path;
@@ -198,15 +234,15 @@ NSString * const c_defaultRealmFileName = @"default.realm";
 
         try {
             if (readonly) {
-                _readGroup = make_unique<Group>(path.UTF8String);
+                _readGroup = make_unique<Group>(path.UTF8String, static_cast<const char *>(key.bytes));
                 _group = _readGroup.get();
             }
             else {
-                _writeLogs.reset(tightdb::getWriteLogs(path.UTF8String));
                 _replication.reset(tightdb::makeWriteLogCollector(path.UTF8String));
                 SharedGroup::DurabilityLevel durability = inMemory ? SharedGroup::durability_MemOnly :
                                                                      SharedGroup::durability_Full;
-                _sharedGroup = make_unique<SharedGroup>(*_replication, durability);
+                _sharedGroup = make_unique<SharedGroup>(*_replication, durability,
+                                                        static_cast<const char *>(key.bytes));
             }
         }
         catch (File::PermissionDenied &ex) {
@@ -217,12 +253,6 @@ NSString * const c_defaultRealmFileName = @"default.realm";
         }
         catch (File::AccessError &ex) {
             *error = make_realm_error(RLMErrorFileAccessError, ex);
-        }
-        catch (SharedGroup::PresumablyStaleLockFile &ex) {
-            *error = make_realm_error(RLMErrorStaleLockFile, ex);
-        }
-        catch (SharedGroup::LockFileButNoData &ex) {
-            *error = make_realm_error(RLMErrorLockFileButNoData, ex);
         }
         catch (exception &ex) {
             *error = make_realm_error(RLMErrorFail, ex);
@@ -294,14 +324,28 @@ NSString * const c_defaultRealmFileName = @"default.realm";
                      readOnly:(BOOL)readonly
                         error:(NSError **)outError
 {
-    return [self realmWithPath:path readOnly:readonly inMemory:NO dynamic:NO schema:nil error:outError];
+    return [self realmWithPath:path key:nil readOnly:readonly inMemory:NO dynamic:NO schema:nil error:outError];
 }
 
 + (instancetype)inMemoryRealmWithIdentifier:(NSString *)identifier {
-    return [self realmWithPath:[RLMRealm writeablePathForFile:identifier] readOnly:NO inMemory:YES dynamic:NO schema:nil error:nil];
+    return [self realmWithPath:[RLMRealm writeablePathForFile:identifier] key:nil
+                      readOnly:NO inMemory:YES dynamic:NO schema:nil error:nil];
+}
+
++ (instancetype)encryptedRealmWithPath:(NSString *)path
+                                   key:(NSData *)key
+                              readOnly:(BOOL)readonly
+                                 error:(NSError **)error
+{
+    if (!key) {
+        @throw [NSException exceptionWithName:@"RLMException" reason:@"Encryption key must not be nil" userInfo:nil];
+    }
+
+    return [self realmWithPath:path key:key readOnly:readonly inMemory:NO dynamic:NO schema:nil error:error];
 }
 
 + (instancetype)realmWithPath:(NSString *)path
+                          key:(NSData *)key
                      readOnly:(BOOL)readonly
                      inMemory:(BOOL)inMemory
                       dynamic:(BOOL)dynamic
@@ -342,8 +386,14 @@ NSString * const c_defaultRealmFileName = @"default.realm";
         return realm;
     }
 
+    if (!key) {
+        @synchronized (s_keysPerPath) {
+            key = s_keysPerPath[path];
+        }
+    }
+
     NSError *error = nil;
-    realm = [[RLMRealm alloc] initWithPath:path readOnly:readonly inMemory:inMemory error:&error];
+    realm = [[RLMRealm alloc] initWithPath:path key:key readOnly:readonly inMemory:inMemory error:&error];
     realm->_dynamic = dynamic;
 
     if (error) {
@@ -399,7 +449,7 @@ NSString * const c_defaultRealmFileName = @"default.realm";
                     createTablesInTransaction(realm, [RLMSchema sharedSchema]);
                 }
                 else {
-                    [RLMRealm migrateRealm:realm];
+                    [RLMRealm migrateRealm:realm key:key];
                 }
 
                 RLMRealmCreateAccessors(realm.schema);
@@ -414,6 +464,25 @@ NSString * const c_defaultRealmFileName = @"default.realm";
     }
 
     return realm;
+}
+
++ (void)setEncryptionKey:(NSData *)key forRealmsAtPath:(NSString *)path {
+    if (!key) {
+        @synchronized (s_keysPerPath) {
+            [s_keysPerPath removeObjectForKey:path];
+            return;
+        }
+    }
+
+    if ([key length] != 64) {
+        @throw [NSException exceptionWithName:@"RLMException"
+                                       reason:@"Encryption key must be exactly 64 bytes"
+                                     userInfo:nil];
+    }
+
+    @synchronized (s_keysPerPath) {
+        s_keysPerPath[path] = key;
+    }
 }
 
 + (void)resetRealmState {
@@ -477,7 +546,7 @@ static void CheckReadWrite(RLMRealm *realm, NSString *msg=@"Cannot write to a re
             // begin the read transaction if needed
             [self getOrCreateGroup];
 
-            LangBindHelper::promote_to_write(*_sharedGroup, *_writeLogs);
+            LangBindHelper::promote_to_write(*_sharedGroup);
 
             if (announce) {
                 [self sendNotifications:RLMRealmDidChangeNotification];
@@ -581,7 +650,7 @@ static void CheckReadWrite(RLMRealm *realm, NSString *msg=@"Cannot write to a re
         if (_sharedGroup->has_changed()) { // Throws
             if (_autorefresh) {
                 if (_group) {
-                    LangBindHelper::advance_read(*_sharedGroup, *_writeLogs);
+                    LangBindHelper::advance_read(*_sharedGroup);
                 }
                 [self sendNotifications:RLMRealmDidChangeNotification];
             }
@@ -608,7 +677,7 @@ static void CheckReadWrite(RLMRealm *realm, NSString *msg=@"Cannot write to a re
         // advance transaction if database has changed
         if (_sharedGroup->has_changed()) { // Throws
             if (_group) {
-                LangBindHelper::advance_read(*_sharedGroup, *_writeLogs);
+                LangBindHelper::advance_read(*_sharedGroup);
             }
             else {
                 // Create the read transaction
@@ -708,18 +777,37 @@ static void CheckReadWrite(RLMRealm *realm, NSString *msg=@"Cannot write to a re
 }
 
 + (NSError *)migrateRealmAtPath:(NSString *)realmPath {
+    NSData *key;
+    @synchronized (s_keysPerPath) {
+        key = s_keysPerPath[realmPath];
+    }
+
     NSError *error;
-    RLMRealm *realm = [RLMRealm realmWithPath:realmPath readOnly:NO inMemory:NO dynamic:YES schema:nil error:&error];
+    RLMRealm *realm = [RLMRealm realmWithPath:realmPath key:key readOnly:NO inMemory:NO dynamic:YES schema:nil error:&error];
     if (error) {
         return error;
     }
 
-    return [self migrateRealm:realm];
+    return [self migrateRealm:realm key:key];
 }
 
-+ (NSError *)migrateRealm:(RLMRealm *)realm {
++ (NSError *)migrateEncryptedRealmAtPath:(NSString *)realmPath key:(NSData *)key {
+    if (!key) {
+        @throw [NSException exceptionWithName:@"RLMException" reason:@"Encryption key must not be nil" userInfo:nil];
+    }
+
     NSError *error;
-    RLMMigration *migration = [RLMMigration migrationForRealm:realm error:&error];
+    RLMRealm *realm = [self realmWithPath:realmPath key:key readOnly:NO inMemory:NO dynamic:YES schema:nil error:&error];
+    if (error) {
+        return error;
+    }
+
+    return [self migrateRealm:realm key:key];
+}
+
++ (NSError *)migrateRealm:(RLMRealm *)realm key:(NSData *)key {
+    NSError *error;
+    RLMMigration *migration = [RLMMigration migrationForRealm:realm key:key error:&error];
     if (error) {
         return error;
     }