+++ tools/RealmBrowser/RealmBrowser/Classes/Realm_Private.h
@@ -21,6 +21,7 @@
 @interface RLMRealm (Dynamic)
 
 + (instancetype)realmWithPath:(NSString *)path
+                          key:(NSData *)key
                      readOnly:(BOOL)readonly
                      inMemory:(BOOL)inMemory
                       dynamic:(BOOL)dynamic