+++ CHANGELOG.md
@@ -8,6 +8,8 @@ x.x.x Release notes (yyyy-MM-dd)
 ### Enhancements
 
 * Add support for encrypting Realm files on disk.
+* Support using KVC-compliant objects without getters or with custom getter
+  names to initialize RLMObjects with `createObjectInRealm` and friends.
 
 ### Bugfixes
 
@@ -18,7 +20,6 @@ x.x.x Release notes (yyyy-MM-dd)
   initialize other objects.
 * Fix a major performance regression on queries on string properties.
 * Fix a memory leak when circularly linked objects are added to a Realm.
-
 
 0.88.0 Release notes (2014-12-02)
 =============================================================